+++ CHANGELOG.md
@@ -16,12 +16,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/dd987e
+- update lock file(2026-05-12)
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+- temporarily switch off not_exactly_x_items assets check
 
 ### Security